requirements.txt
python-decouple
gspread
oauth2client
PyYAML
ollama
rendercv[full]
python-slugify
openai
